feat: add dashboard page
